requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
av==13.0.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grep-ast==0.3.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==7.2.1
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.44.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
openai==1.42.0
opencv-python==4.10.0.84
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.13
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scipy==1.13.1
setuptools==75.1.0
setuptools-scm==8.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.5
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.9.4
zipp==3.20.1
